Input/Output

Functions are provided to read and write data in various formats.

Readers

In addition to reading EEG data the following file reading functions are available.

Neuroimaging.import_biosemi — Function

Import Biosemi files

Neuroimaging.read_elp — Function

Read elp file containing sensor locations

Input

fname: Name or path for the sfp file

coordinate: Coordinate system for electrode location

r: Radius for converting spherical coords

Output

elecs: Array of electrode objects

Neuroimaging.read_avr — Function

Read AVR (.avr) file

Input

fname: Name or path for the AVR file

Output

data: Array of data read from AVR file. Each column represents a channel, and each row represents a point.

chanNames: Channel Names

Neuroimaging.read_evt — Function

Read *.evt file and convert to form for EEG.jl

Neuroimaging.read_sfp — Function

Read sfp file containing sensor locations

Input

fname: Path for the sfp file

Output

elec: Electrodes object

Neuroimaging.read_dat — Function

Read dat files

Arguments

fname: Name or path for the dat file

Returns

x: Range of x values

y: Range of y values

z: Range of z values

complete_data: Array (x × y × z x t)

sample_times

References

File specs were taken from fieldtrip

Neuroimaging.read_bsa — Function

Read Besa's BSA (.bsa) file

Input

fname: Name or path for the BSA file

Output

bsa: Dipole object

Writers

In addition to reading EEG data the following file reading functions are available.

Neuroimaging.write_avr — Function

Write AVR file

Neuroimaging.write_dat — Function

Write dat file