Library

Module

Missing docstring.

Missing docstring for EEG. Check Documenter's build log for details.

Preprocessing

EEG.highpass_filter — Function

High pass filter applied in forward and reverse direction

Simply a wrapper for the DSP.jl functions

Arguments

signals: Signal data in the format samples x channels

cutOff: Cut off frequency in Hz

fs: Sampling rate

order: Filter orde

Returns

filtered signal

filter used on signal

highpass_filter(a::SSR; cutOff::Real=2, fs::Real=samplingrate(a), order::Int=3, tolerance::Real=0.01, kwargs...)

Applly a high pass filter.

A zero phase high pass filter is applied to the data using filtfilt. A check is performed to ensure the filter does not affect the modulation rate. The filter coefficents are stored in the processing field.

Example

a = highpass_filter(a)
# or
a = highpass_filter(a, cutOff = 1)

EEG.lowpass_filter — Function

Low pass filter applied in forward and reverse direction

Simply a wrapper for the DSP.jl functions

Input

signals: Signal data in the format samples x channels

cutOff: Cut off frequency in Hz

fs: Sampling rate

order: Filter orde

Output

filtered signal

filter used on signal

EEG.bandpass_filter — Function

Band pass filter

EEG.compensate_for_filter — Function

Recover the spectrum of signal by compensating for filtering done.

Arguments

filter: The filter used on the spectrum

spectrum: Spectrum of signal

frequencies: Array of frequencies you want to apply the compensation to

fs: Sampling rate

Returns

Spectrum of the signal after comensating for the filter

TODO

Extend this to arbitrary number of dimensions rather than the hard coded 3

EEG.epoch_rejection — Function

Reject epochs based on the maximum peak to peak voltage within an epoch across all channels

Arguments

epochs: Array containing the epoch data in the format samples x epochs x channels

retain_percentage: The percentage of epochs to retain

rejection_method: Method to be used for epoch rejection (peak2peak)

Returns

An array with a reduced amount of entries in the epochs dimension

EEG.peak2peak — Function

Find the peak to peak value for each epoch to be returned to epoch_rejection()

EEG.channel_rejection — Function

Reject channels with too great a variance.

Rejection can be based on a threshold or dynamicly chosen based on the variation of all channels.

Arguments

signals: Array of data in format samples x channels

threshold_abs: Absolute threshold to remove channels with variance above this value

threshold_std: Reject channels with a variance more than n times the std of all channels

Returns

An array indicating the channels to be kept

EEG.remove_template — Function

Remove a template signal from each column of an array

Arguments

signals: Original signals to be modified (samples x channels)

template: Template to remove from each signal

Returns

Signals with template removed

EEG.rereference — Function

Re reference a signals to specific signal channel by index.

If multiple channels are specififed, their average is used as the reference.

Arguments

signals: Original signals to be modified

refChan: Index of channels to be used as reference

Returns

Rereferenced signals

Re-reference a signals to specific signal channel by name.

If multiple channels are specififed, their average is used as the reference. Or you can specify to use the average reference.

Arguments

signals: Original signals to be modified

refChan: List of channels to be used as reference or average

chanNames: List of channel names associated with signals array

Returns

Rereferenced signals

rereference(a::SSR, refChan::Union{AbstractString, Array{AbstractString}}; kwargs...)

Reference data to specified channel(s).

Example

a = rereference(a, "Cz")
# or
a = rereference(a, ["P9", "P10"])