Functions

In addition to the function available for processing EEG data, a number of functions are provided specifically for the processing of SSR data

Import

Neuroimaging.read_EEG — Function

read_EEG(fname::AbstractString)
read_EEG(args...)

Read a file or IO stream and store the data in an GeneralEEG type.

Arguments

fname: Name of the file to be read

min_epoch_length: Minimum epoch length in samples. Shorter epochs will be removed (0)

max_epoch_length: Maximum epoch length in samples. Longer epochs will be removed (0 = all)

valid_triggers: Triggers that are considered valid, others are removed ([1,2])

stimulation_amplitude: Amplitude of stimulation (NaN)

remove_first: Number of epochs to be removed from start of recording (0)

max_epochs: Maximum number of epochs to retain (0 = all)

Supported file formats

BIOSEMI (.bdf)

Querying data

Neuroimaging.samplingrate — Method

samplingrate(t::Type, s::EEG)
samplingrate(s::EEG)

Return the sampling rate of an EEG type in Hz as the requested type. If no type is provided, the sampling rate is returned as a floating point number.

Examples

s = read_EEG(filename)
samplingrate(s)

Neuroimaging.channelnames — Method

channelnames(s::EEG)

Return the names of sensors in EEG measurement.

Examples

s = read_EEG(filename)
channelnames(s)

Manipulating data

Base.hcat — Method

hcat(a::EEG, b::EEG)

Concatenate two EEG measurements together, effectively creating a single long measurement.

Examples

hcat(a, b)

Neuroimaging.add_channel — Method

add_channel(a::EEG, data::Vector, chanLabel::AbstractString)

Add a channel to the EEG type with specified channel names.

Examples

s = read_EEG(filename)
new_channel = mean(s.data, 2)
s = add_channel(s, new_channel, "MeanChannelData")

Neuroimaging.remove_channel! — Method

remove_channel!(a::EEG, channelname::AbstractString)
remove_channel!(a::EEG, channelnames::Array{AbstractString})
remove_channel!(a::EEG, channelidx::Int)
remove_channel!(a::EEG, channelidxs::Array{Int})

Remove channel(s) from EEG as specifed by channelname or channelidx.

Examples

a = read_EEG(filename)
remove_channel!(a, ["TP8", "Cz"])

Neuroimaging.keep_channel! — Method

keep_channel!(a::EEG, channelname::AbstractString)
keep_channel!(a::EEG, channelnames::Array{AbstractString})
keep_channel!(a::EEG, channelidxs::Array{Int})

Remove all channels except those requested from EEG.

Examples

a = read_EEG(filename)
keep_channel!(a, ["P8", "Cz"])

Neuroimaging.trim_channel — Method

trim_channel(a::EEG, stop::Int; start::Int=1)

Trim EEG recording by removing data after stop specifed samples and optionally before start samples.

Examples

s = trim_channel(s, 8192*300, start=8192)

Neuroimaging.merge_channels — Function

merge_channels(a::EEG, merge_Chans::Array{S}, new_name::S) where {S<:AbstractString}
merge_channels(a::EEG, merge_Chans::S, new_name::S) where {S<:AbstractString}

Average EEG channels listed in merge_Chans and label the averaged channel as new_name.

Examples

s = merge_channels(s, ["P6", "P8"], "P68")

Epoching

Neuroimaging.epoch_rejection — Method

epoch_rejection(a::EEG; retain_percentage::Number = 0.95, kwargs...)

Reject epochs such that retain_percentage is retained.

Plotting

Neuroimaging.plot_timeseries — Method

plot_timeseries(s::EEG; channels, fs, kwargs)

Plot an EEG recording.

Plot detailed single channel or general multichanel figure depending on how many channels are requested.

Input

s: EEG type

channels: The channels you want to plot, all if not specified

fs: Sample rate

Other optional arguements are passed to the Plots.jl functions

Output

Returns a figure

Example

plot1 = plot_timeseries(s, channels=["P6", "Cz"], plot_points=8192*4)
draw(PDF("timeseries.pdf", 10inch, 6inch), plot1)