Library

Module

Missing docstring.

Missing docstring for EEG. Check Documenter's build log for details.

To be sorted

EEG.remove_template — Function

Remove a template signal from each column of an array

Arguments

signals: Original signals to be modified (samples x channels)

template: Template to remove from each signal

Returns

Signals with template removed

EEG.import_biosemi — Function

Import Biosemi files

EEG.add_channel — Function

Add a channel to the SSR type with specified channel names.

Example

Add a channel called Merged

s = read_SSR(filename)
new_channel = mean(s.data, 2)
s = add_channel(s, new_channel, "Merged")

EEG.join_triggers — Function

Append the trigger information of one SSR type to another. Places the trigger information at the end of first file

Example

join_triggers(a, b)

EEG.keep_channel! — Function

Remove all channels except those requested from SSR.

Example

Remove all channels except Cz and those in the set called EEG_Vanvooren_2014_Right

a = read_SSR(filename)
keep_channel!(a, [EEG_Vanvooren_2014_Right, "Cz"])

EEG.validate_triggers — Function

Validate trigger channel

EEG.clean_triggers — Function

Clean trigger channel

EEG.remove_channel! — Function

Remove specified channels from SSR.

Example

Remove channel Cz and those in the set called EEG_Vanvooren_2014_Right

a = read_SSR(filename)
remove_channel!(a, [EEG_Vanvooren_2014_Right, "Cz"])

EEG.samplingrate — Function

Return the sampling rate of a steady state type. If no type is provided, the sampling rate is returned as a floating point.

Example

Return the sampling rate of a recording

s = read_SSR(filename)
samplingrate(s)

EEG.Dipole — Type

Dipole type.

Parameters

coord_system: The coordinate system that the locations are stored in

x,y,z: Location of dipole

x,y,z/ori: Orientation of dipole

color: Color of dipole for plotting

state: State of dipol

size: size of dipole

EEG.extract_epochs — Function

Extract epoch data from array of channels.

Input

Array of raw data. Samples x Channels

Dictionary of trigger information

Vector of valid trigger numbers

Number of first triggers to remove

Number of end triggers to remove

Example

epochs = extract_epochs(data, triggers, [1,2], 0, 0)

Extract epoch data from SSR

Arguments

a: A SSR object

valid_triggers: Trigger numbers that are considered valid ([1,2])

remove_first: Remove the first n triggers (0)

remove_last: Remove the last n triggers (0)

Example

epochs = extract_epochs(SSR, valid_triggers=[1,2])

EEG.VolumeImage — Type

Volume Image

This composite type contains volume image information

Fields

data: contains the recorded data

x, y, z, t Arrays containing spatial and time information

method AbstractString of method used to compute tomography

info: additional information in dictionary

processing Fields

The following standard names are used when saving data to the info dictionary.

Regularisation: Regularisation used in tomography

NormalisationConstant: Value used to normalise image to maximum of 1

FileName: Name of file

EEG.find_dipoles — Function

Find all dipole in an activity map.

Determines the local maxima in a 3 dimensional array

Input

s: Activity in 3d matrix

window: Windowing to use in each dimension for min max filter

x,y,z: Coordinates associated with s matrix

Output

dips: An array of dipoles

EEG.find_location — Function

Find index of location of coordinate or dipole in leadfield

Find indicies of location in VolumeImage

IO

EEG.read_elp — Function

Read elp file containing sensor locations

Input

fname: Name or path for the sfp file

coordinate: Coordinate system for electrode location

r: Radius for converting spherical coords

Output

elecs: Array of electrode objects

EEG.write_avr — Function

Write AVR file

EEG.read_avr — Function

Read AVR (.avr) file

Input

fname: Name or path for the AVR file

Output

data: Array of data read from AVR file. Each column represents a channel, and each row represents a point.

chanNames: Channel Names

EEG.read_evt — Function

Read *.evt file and convert to form for EEG.jl

EEG.read_sfp — Function

Read sfp file containing sensor locations

Input

fname: Path for the sfp file

Output

elec: Electrodes object

EEG.read_dat — Function

Read dat files

Arguments

fname: Name or path for the dat file

Returns

x: Range of x values

y: Range of y values

z: Range of z values

complete_data: Array (x × y × z x t)

sample_times

References

File specs were taken from fieldtrip

EEG.read_bsa — Function

Read Besa's BSA (.bsa) file

Input

fname: Name or path for the BSA file

Output

bsa: Dipole object

Channels

EEG.match_sensors — Function

Match a set of electrodes to those provided

usage: lf, valid = matchsensors(electrodes, sensorlabels)

EEG.trim_channel — Function

Trim SSR recording by removing data after stop specifed samples.

Optional Parameters

start Remove samples before this value

Example

Remove the first 8192 samples and everything after 8192*300 samples

s = trim_channel(s, 8192*300, start=8192)

EEG.channelnames — Function

Return the names of sensors in SSR measurement.

Example

s = read_SSR(filename)
channelnames(s)

Change the names of sensors in SSR measurement.

Example

s = read_SSR(filename)
channelnames(s, 1, "Fp1")

EEG.merge_channels — Function

Merge SSR channels listed in merge_Chans and label the averaged channel as new_name

If multiple channels are listed then the average of those channels will be added.

Example

s = merge_channels(s, ["P6", "P8"], "P68")

EEG.channel_rejection — Function

Reject channels with too great a variance.

Rejection can be based on a threshold or dynamicly chosen based on the variation of all channels.

Arguments

signals: Array of data in format samples x channels

threshold_abs: Absolute threshold to remove channels with variance above this value

threshold_std: Reject channels with a variance more than n times the std of all channels

Returns

An array indicating the channels to be kept

Preprocessing

EEG.rereference — Function

Re reference a signals to specific signal channel by index.

If multiple channels are specififed, their average is used as the reference.

Arguments

signals: Original signals to be modified

refChan: Index of channels to be used as reference

Returns

Rereferenced signals

Re-reference a signals to specific signal channel by name.

If multiple channels are specififed, their average is used as the reference. Or you can specify to use the average reference.

Arguments

signals: Original signals to be modified

refChan: List of channels to be used as reference or average

chanNames: List of channel names associated with signals array

Returns

Rereferenced signals

rereference(a::SSR, refChan::Union{AbstractString, Array{AbstractString}}; kwargs...)

Reference data to specified channel(s).

Example

a = rereference(a, "Cz")
# or
a = rereference(a, ["P9", "P10"])

EEG.highpass_filter — Function

High pass filter applied in forward and reverse direction

Simply a wrapper for the DSP.jl functions

Arguments

signals: Signal data in the format samples x channels

cutOff: Cut off frequency in Hz

fs: Sampling rate

order: Filter orde

Returns

filtered signal

filter used on signal

highpass_filter(a::SSR; cutOff::Real=2, fs::Real=samplingrate(a), order::Int=3, tolerance::Real=0.01, kwargs...)

Applly a high pass filter.

A zero phase high pass filter is applied to the data using filtfilt. A check is performed to ensure the filter does not affect the modulation rate. The filter coefficents are stored in the processing field.

Example

a = highpass_filter(a)
# or
a = highpass_filter(a, cutOff = 1)

EEG.lowpass_filter — Function

Low pass filter applied in forward and reverse direction

Simply a wrapper for the DSP.jl functions

Input

signals: Signal data in the format samples x channels

cutOff: Cut off frequency in Hz

fs: Sampling rate

order: Filter orde

Output

filtered signal

filter used on signal

EEG.bandpass_filter — Function

Band pass filter

EEG.compensate_for_filter — Function

Recover the spectrum of signal by compensating for filtering done.

Arguments

filter: The filter used on the spectrum

spectrum: Spectrum of signal

frequencies: Array of frequencies you want to apply the compensation to

fs: Sampling rate

Returns

Spectrum of the signal after comensating for the filter

TODO

Extend this to arbitrary number of dimensions rather than the hard coded 3

EEG.epoch_rejection — Function

Reject epochs based on the maximum peak to peak voltage within an epoch across all channels

Arguments

epochs: Array containing the epoch data in the format samples x epochs x channels

retain_percentage: The percentage of epochs to retain

rejection_method: Method to be used for epoch rejection (peak2peak)

Returns

An array with a reduced amount of entries in the epochs dimension

EEG.peak2peak — Function

Find the peak to peak value for each epoch to be returned to epoch_rejection()

Statistics

EEG.ftest — Function

Calculates the F test as is commonly implemented in SSR research. TODO: Add references to MASTER and Luts et al

Parameters

Sweep measurements. Samples x Sweeps x Channels

Frequency(ies) of interest (Hz)

Sampling rate (Hz)

The amount of data to use on each side of frequency of interest to estimate noise (Hz)

Filter used on the sweep data. If provided then is compensated for

The number of bins to ignore on each side of the frequency of interest

Returns

Signal to noise ratio in dB

Signal phase at frequency of interest

Signal power at frequency of interest

Noise power estimated of side frequencies

F statistic

Plotting

EEG.plot_single_channel_timeseries — Function

Plot a single channel time series

Input

signal: Vector of data

fs: Sample rate

channels: Name of channel to plot

plot_points: Number of points to plot, they will be equally spread. Used to speed up plotting

Other optional arguements are passed to gadfly plot function

Output

Returns a figure

EEG.plot_multi_channel_timeseries — Function

Plot a multi channel time series

Input

signals: Array of data

fs: Sample rate

channels: Name of channels

plot_points: Number of points to plot, they will be equally spread. Used to speed up plotting

Other optional arguements are passed to gadfly plot function

Output

Returns a figure