Types

A feature of the Julia programming language is the strong type system. This package exploits that strength and defines various types for storing information about your neuroimaging data.

A number of types are provided to handle different types of data. Functions are provided to perform common operations on each type. For example, the function channelnames would return the correct information when called on a steady state response or evoked potential data type (yet to be implemented).

Types also exist for storing metadata. For example, electrodes are a sub type of the Sensor type. And the position of the sensors may be in the Talairach space, which is a subtype of the Coordinate type.

A brief description of each type is provided below. See the following documentation sections for more details of each type, including examples and function APIs.

Available types

Measurement

Neuroimaging.SSR — Type

Type for storing steady state response (SSR) data.

The minimum required information for SSR recording is stored in the type. Additional information can be stored in the processing field. To facilitate processing, specific names are used in the processing dictionary.

Fields

data: contains the recorded data

trigers: contains information about timing for creation of epochs

system_codes: contains system information

samplingrate: the sampling rate of the data

modulationrate: the modulation rate of the stimulus

reference_channel: the channel the data has been referenced to

file_path and file_name: where the file was read in from

channel_names: the names of the channels

processing: dictionary type to store analysis

header: additional information read from the file

Additional processing fields

The following standard names are used when saving data to the processing dictionary.

Name: The identifier for the participant

Side: Side of stimulation

Carrier_Frequency: Carrier frequency of the stimulus

Amplitude: Amplitude of the stimulus

epochs: The epochs extracted from the recording

sweeps: The extracted sweeps from the recording

Example

Put an example here

s = SSR("filename")

Sensor

Missing docstring.

Missing docstring for Sensor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrode. Check Documenter's build log for details.

Coordinate

Missing docstring.

Missing docstring for Coordinate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BrainVision. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Talairach. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPM. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UnknownCoordinate. Check Documenter's build log for details.

Other

Missing docstring.

Missing docstring for VolumeImage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Dipole. Check Documenter's build log for details.

To be implemented

Generic electromagnetic type.

Generic EEG type. Then make SSR inherit from this.

Generic MEG type.

Add evoked response type

Add optodes

Add MEG sensors